Adding new MPI tests to code coverage
