+++ tox.ini
@@ -33,7 +33,8 @@ allowlist_externals =
     /usr/bin/mpiexec
 commands =
     pip install -e .[memprof,mpi,test]
-    mpiexec -n 2 --use-hwthread-cpus coverage run --rcfile=.coveragerc -m pytest --cov=armi --cov-config=.coveragerc --cov-report=lcov --ignore=venv --cov-fail-under=80 armi/tests/test_mpiFeatures.py
+    mpiexec -n 2 --use-hwthread-cpus coverage run --rcfile=.coveragerc -m pytest --cov=armi --cov-config=.coveragerc --cov-report=lcov --ignore=venv armi/tests/test_mpiFeatures.py
+    mpiexec -n 2 --use-hwthread-cpus coverage run --rcfile=.coveragerc -m pytest --cov=armi --cov-config=.coveragerc --cov-report=lcov --cov-append --ignore=venv armi/tests/test_mpiFeatures.py
 
 # Second, run code coverage over the rest of the unit tests, and combine the coverage results together
 [testenv:cov2]
@@ -56,6 +57,7 @@ allowlist_externals =
 commands =
     pip install -e .[memprof,mpi,test]
     mpiexec -n 2 --use-hwthread-cpus pytest armi/tests/test_mpiFeatures.py
+    mpiexec -n 2 --use-hwthread-cpus pytest armi/tests/test_mpiParameters.py
 
 [testenv:lint]
 deps=
